src/requirements.txt
langchain
langchain-community
langchain-core

transformers
huggingface-hub
torch

fastapi
uvicorn
python-dotenv
requests
pydantic
